Pipeline architecture / SQLite storage fixes

+++ src/LinqInfer/Learning/Classification/IFloatingPointClassifier.cs
@@ -1,6 +1,6 @@
 ﻿namespace LinqInfer.Learning.Classification
 {
-    internal interface IFloatingPointClassifier<T> : IClassifier<T, float>
+    internal interface IFloatingPointClassifier<T> : IClassifier<T, double>
     {
     }
 }
