+++ src/LinqInfer/Learning/ObjectVector.cs
@@ -2,7 +2,7 @@
 {
     public class ObjectVector<T>
     {
-        public ObjectVector(T value, float[] attributes)
+        public ObjectVector(T value, double[] attributes)
         {
             Attributes = attributes;
             Value = value;
@@ -10,6 +10,6 @@
 
         public T Value { get; private set; }
 
-        public float[] Attributes { get; private set; }
+        public double[] Attributes { get; private set; }
     }
 }
